# openapi.yml
...
openapi: 3.1.0
info:
  title: Postmill API
  version: 3.4.0
  description: |
    Comprehensive OpenAPI description of the **Postmill** backend (NestJS) as it exists in
    `postmill-ai/postmill-app`.

    Postmill schedules social-media and chat posts to 28+ channels, with a calendar, analytics, team
    management, a media library, and — in this fork — a pluggable multi-provider **AI layer**,
    **Analytics v2** persisted dashboards, **social comments** sync/moderation, admin-configurable
    **channel OAuth credentials**, and an OAuth2 authorization-code provider.

    ## Surfaces & authentication

    There are three distinct authentication surfaces:

    1. **Dashboard API** (the bulk of routes). Requires a Postmill user session JWT supplied either as
       the `auth` **cookie** or the `auth` **request header**. The active organization is selected
       with the `showorg` cookie/header. Many routes are further gated by RBAC policies
       (`@CheckPolicies`) or by `isSuperAdmin`.
    2. **Public integration API** (`/public/v1/*`). Requires an organization **API key** (or a
       `pos_`-prefixed OAuth access token) in the `Authorization` header. Intended for
       n8n/Zapier/CLI/3rd-party automation.
    3. **Unauthenticated routes** — registration/login, the OAuth2 authorization/token endpoints,
       Stripe & enterprise webhooks, the social-connect OAuth callback, health/monitor, and public
...

...
    > global route prefix — controller paths are mounted at the root.

...

...
servers:
...
security:
  - cookieAuth: []
  - authHeader: []
paths:
...
  /posts:
    get:
      tags: [Posts]
      summary: Calendar posts within a date range
      operationId: getPosts
      parameters:
        - { name: startDate, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: endDate, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: customer, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Minified posts for the calendar
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/CalendarPost' } }
    post:
      tags: [Posts]
      summary: Create / schedule / draft a post group
      operationId: createPost
      description: |
        Server-side validates every channel's content & settings. Non-draft posts that fail
        validation return a structured `400` (see `PostValidationError`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePostDto' }
      responses:
        '200':
          description: Created post group
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatePostResult' }
        '400': { $ref: '#/components/responses/PostValidation' }
        '406': { $ref: '#/components/responses/SubscriptionRequired' }
...
  /posts/{id}:
    get:
      tags: [Posts]
      summary: Get a single post
      operationId: getPost
      parameters: [{ $ref: '#/components/parameters/PathUuid' }]
      responses: { '200': { description: Post, content: { application/json: { schema: { type: object, additionalProperties: true } } } } }
...
  /posts/{group}:
    delete:
      tags: [Posts]
      summary: Delete a post group
      operationId: deletePost
      parameters: [{ $ref: '#/components/parameters/PathGroup' }]
      responses: { '200': { description: Deleted } }

  # ===================================================================== SOCIAL COMMENTS (fork)
...
  /analytics/v2/posts:
    get:
      tags: [Analytics V2]
      summary: Post-level analytics, sortable & paginated
      operationId: analyticsPosts
      parameters:
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - { $ref: '#/components/parameters/IntegrationsCsv' }
        - { name: sort, in: query, schema: { type: string } }
        - { name: dir, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 25 } }
      responses: { '200': { description: Paginated post analytics } }
  /analytics/v2/post/{postId}:
    get:
      tags: [Analytics V2]
      summary: Single-post analytics detail (KPI header)
      operationId: analyticsPostDetail
      parameters:
        - { name: postId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: date, in: query, schema: { type: string } }
      responses: { '200': { description: Post detail with live fallback for un-snapshotted posts } }
...
  /webhooks/send:
    post:
      tags: [Webhooks]
      summary: Test-send a webhook payload
      operationId: sendWebhook
      parameters: [{ name: url, in: query, required: true, schema: { type: string, format: uri } }]
      requestBody:
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses: { '200': { description: Sent, content: { application/json: { schema: { type: object, properties: { send: { type: boolean } } } } } } }

  # ===================================================================== AUTOPOST
  /autopost:
    get:
      tags: [Autopost]
      summary: List autopost rules
      operationId: getAutoposts
      responses: { '200': { description: Autoposts } }
    post:
      tags: [Autopost]
      summary: Create an autopost rule
      operationId: createAutopost
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AutopostDto' } } }
      responses: { '200': { description: Created } }
  /autopost/{id}:
    put:
      tags: [Autopost]
      summary: Update an autopost rule
      operationId: updateAutopost
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AutopostDto' } } }
      responses: { '200': { description: Updated } }
    delete:
      tags: [Autopost]
      summary: Delete an autopost rule
      operationId: deleteAutopost
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses: { '200': { description: Deleted } }
  /autopost/{id}/active:
    post:
      tags: [Autopost]
      summary: Toggle an autopost rule active
      operationId: changeAutopostActive
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, properties: { active: { type: boolean } } } } }
      responses: { '200': { description: Toggled } }
  /autopost/send:
    post:
      tags: [Autopost]
      summary: Load and parse an RSS/XML feed
      operationId: autopostSend
      parameters: [{ name: url, in: query, required: true, schema: { type: string, format: uri } }]
      responses: { '200': { description: Parsed feed } }

  # ===================================================================== SIGNATURES
...
  /public/posts/{id}:
    get:
      tags: [Public]
      summary: Public preview of a post thread
      operationId: publicPostPreview
      security: []
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses: { '200': { description: Post thread (sanitized integration info), content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } } }
...
  /public/v1/posts:
    get:
      tags: [Public API]
      summary: List posts in a date range
      operationId: publicApiGetPosts
      security: [{ apiKey: [] }]
      parameters:
        - { name: startDate, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: endDate, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: customer, in: query, schema: { type: string } }
      responses: { '200': { description: Posts, content: { application/json: { schema: { type: object, properties: { posts: { type: array, items: { type: object, additionalProperties: true } } } } } } } }
    post:
      tags: [Public API]
      summary: Create a post
      operationId: publicApiCreatePost
      security: [{ apiKey: [] }]
      description: |
        Same validation as the dashboard. Supports `creationMethod` `CLI`|`API` (defaults to `API`).
        If `RESTRICT_UPLOAD_DOMAINS` is set, all media must be hosted on that domain.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePostDto' } } }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/CreatePostResult' } } } }
        '400': { $ref: '#/components/responses/PostValidation' }
  /public/v1/posts/{id}:
    delete:
      tags: [Public API]
      summary: Delete a post by id (deletes its group)
      operationId: publicApiDeletePost
      security: [{ apiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses: { '200': { description: Deleted } }
...
  /public/v1/analytics/post/{postId}:
    get:
      tags: [Public API]
      summary: Legacy single-post analytics
      operationId: publicApiPostAnalytics
      security: [{ apiKey: [] }]
      parameters:
        - { name: postId, in: path, required: true, schema: { type: string } }
        - { name: date, in: query, schema: { type: string, description: Numeric (days) } }
      responses: { '200': { description: Post analytics } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: Session JWT in the `auth` cookie (dashboard API).
    authHeader:
      type: apiKey
      in: header
      name: auth
      description: Session JWT in the `auth` header (alternative to the cookie).
    apiKey:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Organization **API key** or a `pos_`-prefixed OAuth access token, sent **raw** (no `Bearer`
        prefix) in the `Authorization` header. Used by `/public/v1/*`.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
    PathUuid:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PathGroup:
      name: group
      in: path
      required: true
      schema: { type: string }
...
    From:
      name: from
      in: query
      required: true
      schema: { type: string, format: date }
      description: Range start (inclusive). ISO date.
    To:
      name: to
      in: query
      required: true
      schema: { type: string, format: date }
      description: Range end (inclusive). ISO date.
    IntegrationsCsv:
      name: integrations
      in: query
      required: false
      schema: { type: string }
      description: Comma-separated channel ids to filter by.
...
  responses:
...
    SubscriptionRequired:
      description: An active subscription is required for this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    PostValidation:
      description: Per-channel post validation failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PostValidationError' }
  schemas:
...
    Error:
      type: object
      properties:
        statusCode: { type: integer, example: 400 }
        message:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
          example: Bad Request
        error: { type: string, example: Bad Request }
...
    PostValidationError:
      type: object
      description: Structured per-channel validation failure (PostValidationExceptionFilter).
      properties:
        provider: { type: string, example: twitter }
        name: { type: string, example: My Twitter }
        error: { type: string, example: post is too long, please fix it }

    # ---------- Auth DTOs ----------
...
    MediaRef:
      type: object
      properties:
        id: { type: string }
        path: { type: string }
      additionalProperties: true
    PostContent:
      type: object
      required: [content, image]
      properties:
        content: { type: string, description: Sanitized post content }
        id: { type: string }
        delay: { type: integer, description: Delay (ms) for thread items }
        image:
          type: array
          items: { $ref: '#/components/schemas/MediaRef' }
    Post:
      type: object
      required: [integration, value]
      properties:
        type: { type: string }
        integration:
          type: object
          required: [id]
          properties: { id: { type: string } }
        value:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/PostContent' }
        group: { type: string }
        settings:
          type: object
          description: Provider-specific settings (discriminated by `__type`).
          additionalProperties: true
    CreatePostDto:
      type: object
      required: [type, shortLink, date, tags, posts]
      properties:
        type: { type: string, enum: [draft, schedule, now, update] }
        order: { type: string }
        shortLink: { type: boolean }
        inter: { type: number }
        date: { type: string, format: date-time }
        tags:
          type: array
          items:
            type: object
            properties: { value: { type: string }, label: { type: string } }
        posts:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Post' }
      example:
        type: schedule
        shortLink: false
        date: '2026-06-10T14:00:00.000Z'
        tags: []
        posts:
          - integration: { id: clt_integration_123 }
            value:
              - content: 'Hello from the Postmill API 👋'
                image: []
            settings: { __type: twitter }
    CreatePostResult:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        group: { type: string }
...
    CalendarPost:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        content: { type: string }
        publishDate: { type: string, format: date-time }
        state: { type: string, example: QUEUE }
        integration: { type: object, additionalProperties: true }
...
    AutopostDto:
      type: object
      required: [title, onSlot, syncLast, url, active, addPicture, generateContent]
      properties:
        title: { type: string }
        content: { type: string }
        lastUrl: { type: string }
        onSlot: { type: boolean }
        syncLast: { type: boolean }
        url: { type: string, format: uri, description: Public HTTPS URL (SSRF-validated) }
        active: { type: boolean }
        addPicture: { type: boolean }
        generateContent: { type: boolean }
        integrations:
          type: array
          items:
            type: object
            required: [id]
            properties: { id: { type: string } }

    # ---------- Sets / signatures ----------
...